added a check for attempted inversion of ill-conditioned real matrices When inverting an ill-conditioned real matrix, Orpie first computes the matrix condition number using the SVD, and flags an error if it is too large relative to machine precision.  There is no corresponding check for inversion of complex matrices, because the GSL (currently) lacks a complex SVD.
From an efficiency standpoint, SVD is probably not the best choice for a
condition number estimate.  Unfortunately, the GSL doesn't seem to offer any
other options, and I'm not too interested in taking the time to write up
an algorithm myself.  (If I were more motivated, I'd probably try to translate
a routine from LAPACK.)